User for user Sign-up and Login
This component help to loads user login details from the device and
check the server if the details correspond with the server login
details. If not it request the user to update its detail by logining
again . And user can as well sign-up where it post the sign-up details
to the server..

amacoder ltd.